+++ server/lib/notion/blocks/BlockHeadings.tsx
@@ -5,7 +5,7 @@ import TagRegistry from '../../parser/TagRegistry';
 import BlockHandler from '../BlockHandler';
 import getPlainText from '../helpers/getPlainText';
 import { styleWithColors } from '../NotionColors';
-import HandleBlockAnnotations from './utils';
+import HandleBlockAnnotations from './HandleBlockAnnotations';
 
 interface HeadingProps {
   id: string;
