+++ server/lib/parser/DeckParser.ts
@@ -10,7 +10,7 @@ import { SuffixFrom } from "../misc/file";
 import replaceAll from "./helpers/replaceAll";
 import handleClozeDeletions from "./helpers/handleClozeDeletions";
 import sanitizeTags from "../anki/sanitizeTags";
-import preserveNewlinesIfApplicable from "../notion/helpers/perserveNewlinesIfApplicable";
+import preserveNewlinesIfApplicable from "../notion/helpers/preserveNewlinesIfApplicable";
 
 import getYouTubeID from "./helpers/getYouTubeID";
 import getYouTubeEmbedLink from "./helpers/getYouTubeEmbedLink";
